【Fix PIR Unittest No.125、147、481】Fix some 0D uts in PIR mode (part1)
