+++ test/legacy_test/test_zero_dim_sundry_static_api_part3.py
@@ -428,33 +428,6 @@ class TestSundryAPIStatic(unittest.TestCase):
         self.assertEqual(res[4].shape, ())
         self.assertEqual(res[5].shape, ())
 
-    @prog_scope()
-    def test_static_nn_prelu(self):
-        x1 = paddle.full([], 1.0, 'float32')
-        x1.stop_gradient = False
-        out1 = paddle.static.nn.prelu(x1, 'all')
-        grad_list = paddle.static.append_backward(
-            out1.sum(), parameter_list=[x1, out1]
-        )
-        (_, x1_grad), (_, out1_grad) = grad_list
-
-        prog = paddle.static.default_main_program()
-        self.exe.run(paddle.static.default_startup_program())
-        res = self.exe.run(
-            prog,
-            fetch_list=[
-                out1,
-                x1_grad,
-                out1_grad,
-            ],
-        )
-
-        self.assertEqual(res[0].shape, ())
-        self.assertEqual(res[1].shape, ())
-        self.assertEqual(res[2].shape, ())
-        np.testing.assert_allclose(res[0], np.array(1))
-        np.testing.assert_allclose(res[1], np.array(1))
-
     @test_with_pir_api
     @prog_scope()
     def test_while_loop(self):
